% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnect.R
\name{DBconnect}
\alias{DBconnect}
\title{Create a connection to a DBMS}
\usage{
DBconnect(drv, ...)
}
\arguments{
\item{drv}{an object that inherits from [DBIDriver-class],
or an existing [DBIConnection-class]
object (in order to clone an existing connection).}

\item{...}{authentication arguments needed by the DBMS instance; these
typically include `user`, `password`, `host`, `port`, `dbname`, etc.
For details see the appropriate `DBIDriver`.}
}
\description{
Connect to a DBMS going through the appropriate authentication procedure.
Some implementations may allow you to have multiple connections open, so you
may invoke this function repeatedly assigning its output to different
objects.
The authentication mechanism is left unspecified, so check the
documentation of individual drivers for details.
Use [dbCanConnect()] to check if a connection can be established.
}
\seealso{
[dbDisconnect()] to disconnect from a database.
}
\concept{DBIConnector generics}
\concept{DBIDriver generics}
