% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlQuery.R
\name{dbGetQuery}
\alias{dbGetQuery}
\title{Send query, retrieve results and then clear result set}
\usage{
dbGetQuery(conn, statement, ...)
}
\arguments{
\item{conn}{A [DBIConnection-class] object, as returned by
[dbConnect()].}

\item{statement}{a character string containing SQL.}

\item{...}{Other parameters passed on to methods.}
}
\description{
Returns the result of a query as a data frame.
`dbGetQuery()` comes with a default implementation
(which should work with most backends) that calls
[dbSendQuery()], then [dbFetch()], ensuring that
the result is always free-d by [dbClearResult()].
}
\details{
This method is for `SELECT` queries only
(incl. other SQL statements that return a `SELECT`-alike result,
 e. g. execution of a stored procedure).

To execute a stored procedure that does not return a result set,
use [dbExecute()].

Some backends may
support data manipulation statements through this method for compatibility
reasons.  However, callers are strongly advised to use
[dbExecute()] for data manipulation statements.




#' @section Implementation notes:
Subclasses should override this method only if they provide some sort of
performance optimization.
}
\seealso{
For updates: [dbSendStatement()] and [dbExecute()].
}
\concept{DBIConnection generics}
