% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_draw.R
\name{dm_draw}
\alias{dm_draw}
\title{Draw a diagram of the data model}
\usage{
dm_draw(
  dm,
  rankdir = "LR",
  col_attr = "column",
  view_type = "keys_only",
  columnArrows = TRUE,
  graph_attrs = "",
  node_attrs = "",
  edge_attrs = "",
  focus = NULL,
  graph_name = "Data Model"
)
}
\arguments{
\item{dm}{A [`dm`] object.}

\item{rankdir}{Graph attribute for direction (e.g., 'BT' = bottom --> top).}

\item{col_attr}{Deprecated, use `colummn_types` instead.}

\item{view_type}{Can be "keys_only" (default), "all" or "title_only".
It defines the level of details for rendering tables
(only primary and foreign keys, all columns, or no columns).}

\item{columnArrows}{Edges from columns to columns (default: `TRUE`).}

\item{graph_attrs}{Additional graph attributes.}

\item{node_attrs}{Additional node attributes.}

\item{edge_attrs}{Additional edge attributes.}

\item{focus}{A list of parameters for rendering (table filter).}

\item{graph_name}{The name of the graph.}
}
\description{
`dm_draw()` uses \pkg{DiagrammeR} to draw diagrams.
Use [DiagrammeRsvg::export_svg()] to convert the diagram to an SVG file.
}
\seealso{
[dm_set_colors()] for defining the table colors.
}
